underwriting-copilot/softmax-underwriting-copilot/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "softmax-underwriting-copilot"
version = "0.1.0"
description = "Softmax Underwriting Copilot service"
readme = "README.md"
authors = [
  {name = "Softmax", email = "eng@softmax.mn"}
]
requires-python = ">=3.11"
license = {text = "Proprietary"}
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "structlog>=24.1",
  "prometheus-client>=0.19",
  "celery[redis]>=5.3",
  "redis>=5.0",
  "cryptography>=42",
  "python-jose[cryptography]>=3.3",
  "requests>=2.32",
  "httpx>=0.27",
  "orjson>=3.10",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7",
  "pdfplumber>=0.10.0",
  "PyPDF2>=3.0.1",
  "pdfminer.six>=20221105",
  "pymupdf>=1.23.0",
  "Pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "fakeredis>=2.21",
  "freezegun>=1.4"
]

[tool.setuptools.packages.find]
where = ["app"]
include = ["app*"]
